requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Flask-Marshmallow>=0.15.0
marshmallow-sqlalchemy>=0.29.0
mysqlclient>=2.2.0
pymysql>=1.1.0
gunicorn>=21.2.0
PyJWT>=2.8.0
openai>=1.0.0
